@gnarf37 - I'm having some issues here, hoping I might be able to get some help from you at some point as to why this is behaving this way.
Attempting to store an 'embedded' version of all items, to avoid extra queries when loading a build.

The goal here is as follows:

- **_gear** - A document `D3Up_GearSet`, containing references to all gear in the items collection.
- **gear** - A document `D3Up_GearSet_Cache` containing embedded versions of all gear from the items collection.

Everything looks good in terms of configuration and document type, but Epic_Mongo is converting the documents on the `gear` field into references, when it shouldn't be, they should be embedded versions of the items themselves. Only the `_gear` field should be converting the items into references.